% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_aaply.R
\name{fast_aaply}
\alias{fast_aaply}
\title{Fast and more predictable apply function on multidimensional array}
\usage{
fast_aaply(X, MARGIN, FUN)
}
\arguments{
\item{X}{a multidimensional array}

\item{MARGIN}{the dimension for which you want to apply the function on}

\item{FUN}{a function}
}
\description{
The output should be equivalent (check the defination of equivalence in testthat package) to
\code{plyr::aaply(X, MARGIN, FUN, .drop=FALSE)}
}
\examples{
dims <- c(2,10,2,5,1)
arr <- array(1:prod(dims), dims, lapply(dims, function(d) 1:d))
fast_aaply(arr, c(2,4,3), sum)

}
